backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
langchain-openai
langgraph
langchain-core